+++ publish/utils/updateChangeLog.js
@@ -1,50 +1,41 @@
 const fs = require('fs')
-const { jp, formatTime } = require('./index')
-const pkgDir = '../../package.json'
-const pkg = require(pkgDir)
-const version = require('../version.json')
+const { jp } = require('./index')
 const chalk = require('chalk')
-const pkg_bak = JSON.stringify(pkg, null, 2)
-const version_bak = JSON.stringify(version, null, 2)
 const changelogPath = jp('../../CHANGELOG.md')
-const { parseChangelog } = require('./parseChangelog')
+const pkgPath = jp('../../package.json')
+const versionPath = jp('../version.json')
+const { applyReleaseVersion, formatReleaseDate, formatReleaseVersion } = require('../../scripts/release/versioning')
 
-const getPrevVer = () => parseChangelog(fs.readFileSync(changelogPath, 'utf-8').toString()).then(versions => {
-  if (!versions.length) throw new Error('CHANGELOG 无法解析到版本号')
-  return versions[0].version
-})
-
-const updateChangeLog = async(newVerNum, newChangeLog) => {
-  let changeLog = fs.readFileSync(changelogPath, 'utf-8')
-  const prevVer = await getPrevVer()
-  const log = `## [${newVerNum}](${pkg.repository.url.replace(/^git\+(http.+)\.git$/, '$1')}/compare/v${prevVer}...v${newVerNum}) - ${formatTime()}\n\n${newChangeLog}`
-  fs.writeFileSync(changelogPath, changeLog.replace(/(## [?0.1.1]?)/, log + '\n$1'), 'utf-8')
-}
-
+const readJson = filePath => JSON.parse(fs.readFileSync(filePath, 'utf-8'))
 
 module.exports = async newVerNum => {
-  if (!newVerNum) newVerNum = pkg.version
+  const pkg = readJson(pkgPath)
+  const version = readJson(versionPath)
+  const pkg_bak = JSON.stringify(pkg, null, 2)
+  const version_bak = JSON.stringify(version, null, 2)
+  const changelog_bak = fs.readFileSync(changelogPath, 'utf-8')
+  if (!newVerNum) newVerNum = formatReleaseVersion()
   const newMDChangeLog = fs.readFileSync(jp('../changeLog.md'), 'utf-8')
-  version.history.unshift({
-    version: version.version,
-    desc: version.desc,
+  const nextState = applyReleaseVersion({
+    packageJson: pkg,
+    versionJson: version,
+    changelogMarkdown: changelog_bak,
+    releaseNotesMarkdown: newMDChangeLog,
+    version: newVerNum,
+    releaseDate: formatReleaseDate(),
   })
-  version.version = newVerNum
-  version.desc = newMDChangeLog.replace(/(?:^|(\n))#{1,6} (.+)\n/g, '$1$2').trim()
-  pkg.version = newVerNum
-  pkg.versionCode = pkg.versionCode + 1
 
   console.log(chalk.blue('new version: ') + chalk.green(newVerNum))
 
-  fs.writeFileSync(jp('../version.json'), JSON.stringify(version) + '\n', 'utf-8')
+  fs.writeFileSync(versionPath, JSON.stringify(nextState.versionJson) + '\n', 'utf-8')
 
-  fs.writeFileSync(jp(pkgDir), JSON.stringify(pkg, null, 2) + '\n', 'utf-8')
+  fs.writeFileSync(pkgPath, JSON.stringify(nextState.packageJson, null, 2) + '\n', 'utf-8')
 
-  await updateChangeLog(newVerNum, newMDChangeLog)
+  fs.writeFileSync(changelogPath, nextState.changelogMarkdown, 'utf-8')
 
   return {
     pkg_bak,
     version_bak,
+    changelog_bak,
   }
 }
-